src/main/java/ru/netology/javaqa/MoviePoster/service/PosterManager.java
package ru.netology.javaqa.MoviePoster.service;

public class PosterManager {
    private String[] movies = new String[0];
    private int limit;

    public PosterManager() {
        this.limit = 5;
    }

    public PosterManager(int limit) {
        this.limit = limit;
    }

    public void addMovie(String movie) {
        String[] tmp = new String[movies.length + 1];
        for (int i = 0; i < movies.length; i++) {
            tmp[i] = movies[i];
        }
        tmp[tmp.length - 1] = movie;
        movies = tmp;
    }

    public String[] findAll() {
        return movies;
    }

    public String[] findLast() {
        int resultlegth;
        if (movies.length < limit) {
            resultlegth = movies.length;
        } else {
            resultlegth = limit;
        }
        String[] tmp = new String[resultlegth];
        for (int i = 0; i < tmp.length; i++) {
            tmp[i] = movies[movies.length - 1 - i];
        }
        return tmp;
    }

}
